# Conflicts:
#	src/papilo/presolvers/SimplifyInequalities.hpp

+++ src/papilo/misc/Wrappers.hpp
@@ -515,8 +515,7 @@ postsolve( PostsolveStorage<REAL>& postsolveStorage,
    if( !objective_reference.empty() )
    {
       if( origfeas && status == PostsolveStatus::kOk &&
-          postsolveStorage.num.isFeasEq(
-              boost::lexical_cast<double>( objective_reference ), origobj ) )
+          postsolveStorage.num.isFeasEq( boost::lexical_cast<double>( objective_reference ), double(origobj) ) )
          fmt::print( "validation: SUCCESS\n" );
       else
          fmt::print( "validation: FAILURE\n" );
